+++ fluent.runtime/tests/format/test_placeables.py
@@ -10,8 +10,8 @@ from ..utils import dedent_ftl
 
 class TestPlaceables(unittest.TestCase):
     def setUp(self):
-        self.ctx = FluentBundle(['en-US'], use_isolating=False)
-        self.ctx.add_resource(FluentResource(dedent_ftl("""
+        self.bundle = FluentBundle(['en-US'], use_isolating=False)
+        self.bundle.add_resource(FluentResource(dedent_ftl("""
             message = Message
                     .attr = Message Attribute
             -term = Term
@@ -42,46 +42,46 @@ class TestPlaceables(unittest.TestCase):
         """)))
 
     def test_placeable_message(self):
-        val, errs = self.ctx.format('uses-message', {})
+        val, errs = self.bundle.format_pattern(self.bundle.get_message('uses-message').value, {})
         self.assertEqual(val, 'Message')
         self.assertEqual(len(errs), 0)
 
     def test_placeable_message_attr(self):
-        val, errs = self.ctx.format('uses-message-attr', {})
+        val, errs = self.bundle.format_pattern(self.bundle.get_message('uses-message-attr').value, {})
         self.assertEqual(val, 'Message Attribute')
         self.assertEqual(len(errs), 0)
 
     def test_placeable_term(self):
-        val, errs = self.ctx.format('uses-term', {})
+        val, errs = self.bundle.format_pattern(self.bundle.get_message('uses-term').value, {})
         self.assertEqual(val, 'Term')
         self.assertEqual(len(errs), 0)
 
     def test_placeable_bad_message(self):
-        val, errs = self.ctx.format('bad-message-ref', {})
-        self.assertEqual(val, 'Text not-a-message')
+        val, errs = self.bundle.format_pattern(self.bundle.get_message('bad-message-ref').value, {})
+        self.assertEqual(val, 'Text {not-a-message}')
         self.assertEqual(len(errs), 1)
         self.assertEqual(
             errs,
             [FluentReferenceError("Unknown message: not-a-message")])
 
     def test_placeable_bad_message_attr(self):
-        val, errs = self.ctx.format('bad-message-attr-ref', {})
-        self.assertEqual(val, 'Text Message')
+        val, errs = self.bundle.format_pattern(self.bundle.get_message('bad-message-attr-ref').value, {})
+        self.assertEqual(val, 'Text {message.not-an-attr}')
         self.assertEqual(len(errs), 1)
         self.assertEqual(
             errs,
             [FluentReferenceError("Unknown attribute: message.not-an-attr")])
 
     def test_placeable_bad_term(self):
-        val, errs = self.ctx.format('bad-term-ref', {})
-        self.assertEqual(val, 'Text -not-a-term')
+        val, errs = self.bundle.format_pattern(self.bundle.get_message('bad-term-ref').value, {})
+        self.assertEqual(val, 'Text {-not-a-term}')
         self.assertEqual(len(errs), 1)
         self.assertEqual(
             errs,
             [FluentReferenceError("Unknown term: -not-a-term")])
 
     def test_cycle_detection(self):
-        val, errs = self.ctx.format('self-referencing-message', {})
+        val, errs = self.bundle.format_pattern(self.bundle.get_message('self-referencing-message').value, {})
         self.assertEqual(val, 'Text ???')
         self.assertEqual(len(errs), 1)
         self.assertEqual(
@@ -89,7 +89,7 @@ class TestPlaceables(unittest.TestCase):
             [FluentCyclicReferenceError("Cyclic reference")])
 
     def test_mutual_cycle_detection(self):
-        val, errs = self.ctx.format('cyclic-msg1', {})
+        val, errs = self.bundle.format_pattern(self.bundle.get_message('cyclic-msg1').value, {})
         self.assertEqual(val, 'Text1 Text2 ???')
         self.assertEqual(len(errs), 1)
         self.assertEqual(
@@ -97,53 +97,53 @@ class TestPlaceables(unittest.TestCase):
             [FluentCyclicReferenceError("Cyclic reference")])
 
     def test_allowed_self_reference(self):
-        val, errs = self.ctx.format('self-attribute-ref-ok', {})
+        val, errs = self.bundle.format_pattern(self.bundle.get_message('self-attribute-ref-ok').value, {})
         self.assertEqual(val, 'Parent Attribute')
         self.assertEqual(len(errs), 0)
-        val, errs = self.ctx.format('self-parent-ref-ok.attr', {})
+        val, errs = self.bundle.format_pattern(self.bundle.get_message('self-parent-ref-ok').attributes['attr'], {})
         self.assertEqual(val, 'Attribute Parent')
         self.assertEqual(len(errs), 0)
 
 
 class TestSingleElementPattern(unittest.TestCase):
     def test_single_literal_number_isolating(self):
-        self.ctx = FluentBundle(['en-US'], use_isolating=True)
-        self.ctx.add_resource(FluentResource('foo = { 1 }'))
-        val, errs = self.ctx.format('foo')
+        self.bundle = FluentBundle(['en-US'], use_isolating=True)
+        self.bundle.add_resource(FluentResource('foo = { 1 }'))
+        val, errs = self.bundle.format_pattern(self.bundle.get_message('foo').value)
         self.assertEqual(val, '1')
         self.assertEqual(errs, [])
 
     def test_single_literal_number_non_isolating(self):
-        self.ctx = FluentBundle(['en-US'], use_isolating=False)
-        self.ctx.add_resource(FluentResource('foo = { 1 }'))
-        val, errs = self.ctx.format('foo')
+        self.bundle = FluentBundle(['en-US'], use_isolating=False)
+        self.bundle.add_resource(FluentResource('foo = { 1 }'))
+        val, errs = self.bundle.format_pattern(self.bundle.get_message('foo').value)
         self.assertEqual(val, '1')
         self.assertEqual(errs, [])
 
     def test_single_arg_number_isolating(self):
-        self.ctx = FluentBundle(['en-US'], use_isolating=True)
-        self.ctx.add_resource(FluentResource('foo = { $arg }'))
-        val, errs = self.ctx.format('foo', {'arg': 1})
+        self.bundle = FluentBundle(['en-US'], use_isolating=True)
+        self.bundle.add_resource(FluentResource('foo = { $arg }'))
+        val, errs = self.bundle.format_pattern(self.bundle.get_message('foo').value, {'arg': 1})
         self.assertEqual(val, '1')
         self.assertEqual(errs, [])
 
     def test_single_arg_number_non_isolating(self):
-        self.ctx = FluentBundle(['en-US'], use_isolating=False)
-        self.ctx.add_resource(FluentResource('foo = { $arg }'))
-        val, errs = self.ctx.format('foo', {'arg': 1})
+        self.bundle = FluentBundle(['en-US'], use_isolating=False)
+        self.bundle.add_resource(FluentResource('foo = { $arg }'))
+        val, errs = self.bundle.format_pattern(self.bundle.get_message('foo').value, {'arg': 1})
         self.assertEqual(val, '1')
         self.assertEqual(errs, [])
 
     def test_single_arg_missing_isolating(self):
-        self.ctx = FluentBundle(['en-US'], use_isolating=True)
-        self.ctx.add_resource(FluentResource('foo = { $arg }'))
-        val, errs = self.ctx.format('foo')
+        self.bundle = FluentBundle(['en-US'], use_isolating=True)
+        self.bundle.add_resource(FluentResource('foo = { $arg }'))
+        val, errs = self.bundle.format_pattern(self.bundle.get_message('foo').value)
         self.assertEqual(val, 'arg')
         self.assertEqual(len(errs), 1)
 
     def test_single_arg_missing_non_isolating(self):
-        self.ctx = FluentBundle(['en-US'], use_isolating=False)
-        self.ctx.add_resource(FluentResource('foo = { $arg }'))
-        val, errs = self.ctx.format('foo')
+        self.bundle = FluentBundle(['en-US'], use_isolating=False)
+        self.bundle.add_resource(FluentResource('foo = { $arg }'))
+        val, errs = self.bundle.format_pattern(self.bundle.get_message('foo').value)
         self.assertEqual(val, 'arg')
         self.assertEqual(len(errs), 1)